requirements.txt
# Core LangChain & LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.54.0

# Data Processing
pydantic>=2.10.0
pydantic-core>=2.27.0

# Parsing & Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# UI/CLI
streamlit>=1.39.0
typer>=0.15.0

# Development & Testing
pytest>=8.0.0
black>=24.0.0
flake8>=7.0.0
mypy>=1.13.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0

# Logging & Debugging
loguru==0.7.2
